refactor: Use rust arrays rather than ndarray
Using ndarray has some disadvantages, namely that it is difficult to
convert between types of values. So rather than using this to store the
values I am just using Vecs of raw arrays in rust, which is much simpler
to deal with. These have the same layout in memory as the ndarray
structures and have the added benefit of including guarantees on the
dimensionality.