refactor: Move gsd bindings to their own file
I am starting to make many rust abstractions to the C interface, which I
am putting inside the gsd_bindings module. To make it easier to continue
to do this I have extracted the gsd_bindings module to it's own folder.

An additional part of this restructure is defining rust enums for the
GSD types which are converted in an unusual way.