refactor: Create neighbours iterator and refactor order
The neighbours iterator function is one to get the nearest neighbours of
a configuration. It wraps the required functionality of the Rtree
providing a simpler interface.

This also changes the representation of the Position struct which now
includes the cell rather than a reference to it. This is to simplify the
references in the iterator.