feat: Create separate writer thread for output
This creates a thread which is dedicated to writing the csv files.
This now means there is a thread for reading the gsd file, which is the
main thread. A pool of threads for analysing the data. And a thread for
just writing the output.

This architecture, assuming everything is correct, appears to have a
speed up which is close to 5 times what I am getting with a single
thread. A result which is truly amazing.