feat: Include calculation of voronoi area for molecules
This calculates the voronoi area for each of the molecules using the
voronoi implementation in the CGAL library. This was chosen for being
the simplest to get started quickly, since I need it to work now! There
is a large issue with this dependency, since it requires the CGAL to be
installed, however this can be addressed later.