fix: Wrap all particles inside the box before voronoi
Hoomd allows for particles to be outside of the simulation cell when it
outputs particles to a gsd file. This was causing issues with the
calculation of the voronoi areas, since that is expecting all particles
to be inside the simulation cell.

This includes a function call to wrap the positions of particles inside
the simulation cell so that this is no longer an issue.

Fixes #10