feat: Optionally compute area of cells
This will only compute the area when the --voronoi command line flag is
passed, a response to frames having issues with the voronoi
calculations. I am not entirely sure why there are issues, only that
they exist. Since it is easier to only calculate the area for some
trajectories, this is the approach I am taking.